Added Initialize() and Cleanup() for components.
Added some logging messages.

+++ src/Malsys.Common/Parsing/ParserMessagesLogger.cs
@@ -16,7 +16,7 @@ namespace Malsys.Parsing {
 		}
 
 		public override string GetMessageTypeId(ParsserMessageType msgType) {
-			return ((int)msgType).ToString();
+			return msgType.ToString();
 		}
 
 		protected override string resolveMessage(ParsserMessageType msgType, out MessageType type, params object[] args) {
